requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0
pydantic==2.5.0
python-multipart==0.0.6
openai==1.3.5
python-telegram-bot==20.7
celery==5.3.4
redis==5.0.1
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
smtplib-ssl==1.0.4
python-dotenv==1.0.0
pandas==2.1.3
numpy==1.24.3
aiofiles==23.2.1
Pillow==10.1.0
httpx==0.25.2
jinja2==3.1.2
PyPDF2==3.0.1
python-docx==1.1.0
aiohttp==3.9.1
cryptography==41.0.8